refactor company parts
