+++ Unosquare.CodeChallenge/WarehouseManagement/src/app/app.module.ts
@@ -10,17 +10,17 @@ import { ProductsComponent } from './components/products/products.component';
 import { CompaniesComponent } from './components/companies/companies.component';
 import { SearchpageComponent } from './components/searchpage/searchpage.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { AddcompanyComponent } from './components/parts/addcompany/addcompany.component';
+import { AddcompanyComponent } from './components/companies/parts/addcompany/addcompany.component';
 import { MatDialogModule, MatDialogRef } from '@angular/material/dialog';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { SearchcompanyBarComponent } from './components/parts/searchcompany-bar/searchcompany-bar.component';
-import { CompaniesdatagridComponent } from './components/parts/companiesdatagrid/companiesdatagrid.component';
+import { SearchcompanyBarComponent } from './components/companies/parts/searchcompany-bar/searchcompany-bar.component';
+import { CompaniesdatagridComponent } from './components/companies/parts/companiesdatagrid/companiesdatagrid.component';
 import { HttpClientModule } from '@angular/common/http';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { FormsModule } from '@angular/forms';
-import { CompanydeletemodalComponent } from './components/parts/companydeletemodal/companydeletemodal.component';
+import { CompanydeletemodalComponent } from './components/companies/parts/companydeletemodal/companydeletemodal.component';
 import { AddproductComponent } from './components/products/parts/addproduct/addproduct.component';
 import { SearchbarComponent } from './components/products/parts/searchbar/searchbar.component';
 import { ProductdatagridComponent } from './components/products/parts/productdatagrid/productdatagrid.component';
