+++ Unosquare.CodeChallenge/WarehouseManagement/src/app/components/companies/parts/addcompany/addcompany.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { MatDialogModule, MatDialogRef } from '@angular/material/dialog';
-import { CompaniesService } from '../../../services/companies.service';
+import { CompaniesService } from '../../../../services/companies.service';
 
 import { AddcompanyComponent } from './addcompany.component';
