+++ Unosquare.CodeChallenge/WarehouseManagement/src/app/components/companies/parts/searchcompany-bar/searchcompany-bar.component.spec.ts
